Move account control methods into trait and add setter to handle converting the object's value
